Multiple vimrc changes
Finally getting rid of an ancient bug where I was disabling the colour
scheme wrong on Windows, causing error messages to appear because I
disabled the *plugin* and not the *setting* of the colourscheme.

Disabled a tex checker that's been a pain in the ass.

Removed clojure-related plugins.

Also, made the vim install script install from VundleVim instead of
gmarik. Shouldn't matter much in practise, but makes more sense.